move utils to canonical packages, update imports / exports
