+++ apps/iframe/src/utils/checkIfRequestRequiresConfirmation.ts
@@ -1,4 +1,4 @@
-import { HappyMethodNames } from "@happy.tech/common"
+import { HappyMethodNames } from "@happy.tech/wallet-common"
 import type { Msgs, ProviderMsgsFromApp } from "@happy.tech/wallet-common"
 import { requiresApproval } from "@happy.tech/wallet-common"
 import { Permissions } from "#src/constants/permissions"
