+++ apps/submitter/lib/services/index.ts
@@ -1,9 +1,8 @@
-import type { Hash } from "@happy.tech/common"
+import { type Hash, LruCache } from "@happy.tech/common"
 import { db } from "#lib/database"
 import { env } from "#lib/env"
 import type { SimulateOutput } from "#lib/handlers/simulate"
 import { EvmReceiptService } from "#lib/services/EvmReceiptService"
-import { LruCache } from "#lib/utils/LruCache"
 import { BlockService } from "./BlockService"
 import { BoopNonceManager } from "./BoopNonceManager"
 import { BoopReceiptService } from "./BoopReceiptService"
