+++ support/common/lib/utils/address.ts
@@ -10,3 +10,12 @@ import type { Address } from "./types"
 export function isAddress(str: string): str is Address {
     return viemIsAddress(str, { strict: false })
 }
+
+/**
+ * Truncates an address into a shorter representation by displaying the first `digits` characters
+ * and the last `digits` characters.
+ */
+export const shortenAddress = (address?: Address, digits = 5) => {
+    if (!address) return ""
+    return `${address.substring(0, digits + 2)}...${address.substring(address.length - digits)}`
+}
